+++ src/main/java/com/github/abraham054/finalreality/model/character/player/PlayerCharacter.java
@@ -5,7 +5,7 @@ import com.github.abraham054.finalreality.model.character.ICharacter;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.Executors;
-import com.github.abraham054.finalreality.model.weapon.CommonWeapon.CommonWeapon;
+import com.github.abraham054.finalreality.model.character.player.PlayerClasses.PlayerClasses;
 import com.github.abraham054.finalreality.model.weapon.Weapon;
 import org.jetbrains.annotations.NotNull;
 
@@ -16,36 +16,38 @@ import org.jetbrains.annotations.NotNull;
  * @author <Your name>
  */
 public abstract class PlayerCharacter extends AbstractCharacter {
-  protected static Weapon equippedWeapon = null;
+  protected Weapon equippedWeapon = null;
+  protected final PlayerClasses playerClass;
 
   /**
    * Creates a new character.
    *
-   * @param name
-   *     the character's name
-   * @param turnsQueue
-   *     the queue with the characters waiting for their turn
-   * @param defense
-   *     the character's defense
-   * @param healthPoints
-   *     the character's health points
+   * @param name          the character's name
+   * @param turnsQueue    the queue with the characters waiting for their turn
+   * @param defense       the character's defense
+   * @param healthPoints  the character's health points
+   * @param playerClass   the character's class
    */
   protected PlayerCharacter(@NotNull BlockingQueue<ICharacter> turnsQueue,
-                            @NotNull String name, int defense, int healthPoints) {
+                            @NotNull String name, int defense, int healthPoints,PlayerClasses playerClass) {
     super(turnsQueue, name,defense,healthPoints);
+    this.playerClass = playerClass;
   }
 
   /**
-   * The character equips a non magic weapon
+   * Returns the class of the character.
    * */
-  public void equipNormalWeapon(CommonWeapon weapon) {
-    equippedWeapon = weapon;
-  }
+  public PlayerClasses getCharacterClass() { return this.playerClass; }
+
+  /**
+   * The character equips a weapon
+   * */
+  public abstract void equipWeapon(Weapon weapon);
 
   /**
    * Returns the character's equipped weapon
    * */
-  public static Weapon getEquippedWeapon() { return PlayerCharacter.equippedWeapon; }
+  public Weapon getEquippedWeapon() { return this.equippedWeapon; }
 
   /**
    * Sets a scheduled executor to make this character (thread) wait for {@code speed / 10}