+++ src/main/java/com/github/abraham054/finalreality/model/character/player/commonPlayer/CommonPlayer.java
@@ -2,39 +2,25 @@ package com.github.abraham054.finalreality.model.character.player.commonPlayer;
 
 import com.github.abraham054.finalreality.model.character.ICharacter;
 import com.github.abraham054.finalreality.model.character.player.PlayerCharacter;
+import com.github.abraham054.finalreality.model.character.player.PlayerClasses.CommonPlayerClass;
 import org.jetbrains.annotations.NotNull;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 
-
-public class CommonPlayer extends PlayerCharacter {
-    private final CommonPlayerClass commonClass;
+public abstract class CommonPlayer extends PlayerCharacter {
 
     /**
      * Creates a new common character.
      *
-     * @param name
-     *     the character's name
-     * @param turnsQueue
-     *     the queue with the characters waiting for their turn
-     * @param defense
-     *     the character's defense
-     * @param healthPoints
-     *     the character's health points
-     * @param commonClass
-     *     the character's not magic class
+     * @param name          the character's name
+     * @param turnsQueue    the queue with the characters waiting for their turn
+     * @param defense       the character's defense
+     * @param healthPoints  the character's health points
      */
-    protected CommonPlayer(@NotNull String name,
-                        @NotNull BlockingQueue<ICharacter> turnsQueue, int defense, int healthPoints,
-                        CommonPlayerClass commonClass){
-        super(turnsQueue, name,defense,healthPoints);
-        this.commonClass = commonClass;
+    protected CommonPlayer(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue,
+                           int defense, int healthPoints,CommonPlayerClass playerClass){
+        super(turnsQueue, name,defense,healthPoints,playerClass);
     }
-
-    /**
-     * Returns the class of the character.
-     * */
-    public CommonPlayerClass getCharacterClass() { return this.commonClass; }
 
     @Override
     public boolean equals(Object o) {
@@ -45,12 +31,12 @@ public class CommonPlayer extends PlayerCharacter {
             return false;
         }
         final CommonPlayer that = (CommonPlayer) o;
-        return getCharacterClass() == that.getCharacterClass()
-                && getName().equals(that.getName());
+        return that.hashCode() == hashCode();
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getCharacterClass());
+        return Objects.hash(getCharacterClass(),getName());
     }
+
 }