+++ src/main/java/com/github/abraham054/finalreality/model/character/player/magicPlayer/MagicPlayer.java
@@ -2,57 +2,35 @@ package com.github.abraham054.finalreality.model.character.player.magicPlayer;
 
 import com.github.abraham054.finalreality.model.character.ICharacter;
 import com.github.abraham054.finalreality.model.character.player.PlayerCharacter;
-import com.github.abraham054.finalreality.model.weapon.MagicWeapon.MagicWeapon;
+import com.github.abraham054.finalreality.model.character.player.PlayerClasses.MagicPlayerClass;
 import org.jetbrains.annotations.NotNull;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 
-public class MagicPlayer extends PlayerCharacter {
-    private final MagicPlayerClass magicClass;
-    private static int mana;
+public abstract class MagicPlayer extends PlayerCharacter {
+    private int mana;
 
     /**
      * Creates a new character.
      *
-     * @param name
-     *     the character's name
-     * @param turnsQueue
-     *     the queue with the characters waiting for their turn
-     * @param defense
-     *     the character's defense
-     * @param healthPoints
-     *     the character's health points
-     * @param magicClass
-     *     the character's magic class
-     * @param mana
-     *     the character's mana
+     * @param name          the character's name
+     * @param turnsQueue    the queue with the characters waiting for their turn
+     * @param defense       the character's defense
+     * @param healthPoints  the character's health points
+     * @param mana          the character's mana
      */
-    public MagicPlayer(@NotNull String name,
-                       @NotNull BlockingQueue<ICharacter> turnsQueue, int defense, int healthPoints,
-                       MagicPlayerClass magicClass, int mana){
-        super(turnsQueue, name,defense,healthPoints);
-        this.magicClass = magicClass;
-        MagicPlayer.mana = mana;
-    }
-
-    /**
-     * The magic character equips a magic weapon
-     * */
-    public void equipMagicWeapon(MagicWeapon weapon) {
-        equippedWeapon = weapon;
+    public MagicPlayer(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue,
+                       int defense, int healthPoints, int mana,MagicPlayerClass playerClass){
+        super(turnsQueue, name,defense,healthPoints,playerClass);
+        this.mana = mana;
     }
 
     /**
      * Returns the character's mana
      *  */
-    public static int getMana() {
-        return mana;
+    public int getMana() {
+        return this.mana;
     }
-
-    /**
-     * Returns the class of the character.
-     * */
-    public MagicPlayerClass getCharacterClass() { return this.magicClass; }
 
     @Override
     public boolean equals(Object o) {
@@ -63,12 +41,11 @@ public class MagicPlayer extends PlayerCharacter {
             return false;
         }
         final MagicPlayer that = (MagicPlayer) o;
-        return getCharacterClass() == that.getCharacterClass()
-                && getName().equals(that.getName());
+        return that.hashCode() == hashCode();
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getCharacterClass());
+        return Objects.hash(getCharacterClass(),getName());
     }
 }