+++ src/main/java/com/github/abraham054/finalreality/model/weapon/MagicWeapon/MagicWeapon.java
@@ -1,6 +1,8 @@
 package com.github.abraham054.finalreality.model.weapon.MagicWeapon;
 
 import com.github.abraham054.finalreality.model.weapon.Weapon;
+import com.github.abraham054.finalreality.model.weapon.WeaponTypes.MagicWeaponType;
+import java.util.Objects;
 
 public class MagicWeapon extends Weapon {
     private final int magicDamage;
@@ -8,16 +10,11 @@ public class MagicWeapon extends Weapon {
 
     /**
      * Creates a magic weapon.
-     * @param name
-     *     the magic weapon's name
-     * @param weight
-     *     the magic weapon's name
-     * @param damage
-     *     the magic weapon's damage
-     * @param type
-     *     the magic weapon's type
-     * @param magicDamage
-     *     the magic weapon's magic damage
+     * @param name          the magic weapon's name
+     * @param weight        the magic weapon's name
+     * @param damage        the magic weapon's damage
+     * @param type          the magic weapon's type
+     * @param magicDamage   the magic weapon's magic damage
      */
     public MagicWeapon(String name, int weight, int damage, MagicWeaponType type, int magicDamage) {
         super(name,weight,damage);
@@ -35,7 +32,25 @@ public class MagicWeapon extends Weapon {
     /**
      * Returns the magic weapon's type
      * */
+    @Override
     public MagicWeaponType getType(){
         return type;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof MagicWeapon)) {
+            return false;
+        }
+        final MagicWeapon weapon = (MagicWeapon) o;
+        return weapon.hashCode() == hashCode();
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getName(), getWeight(), getDamage(), getType(),getMagicDamage());
     }
 }