+++ src/main/java/com/github/abraham054/finalreality/model/weapon/Weapon.java
@@ -1,6 +1,6 @@
 package com.github.abraham054.finalreality.model.weapon;
 
-import java.util.Objects;
+import com.github.abraham054.finalreality.model.weapon.WeaponTypes.WeaponType;
 
 /**
  * A class that holds all the information of a weapon.
@@ -8,20 +8,19 @@ import java.util.Objects;
  * @author Ignacio Slater Muñoz.
  * @author <Your name>
  */
-public class Weapon {
+public abstract class Weapon {
 
   private final String name;
   private final int weight;
   private final int damage;
 
   /**
-   * Creates a weapon
-   * @param name
-   *     the weapon's name
-   * @param weight
-   *     the weapon's weight
+   * Creates a weapon.
+   * @param name    the weapon's name
+   * @param weight  the weapon's weight
+   * @param damage  the weapon's damage
    */
-  public Weapon(String name,int weight, int damage) {
+  protected Weapon(String name,int weight, int damage) {
     this.name = name;
     this.weight = weight;
     this.damage = damage;
@@ -30,7 +29,7 @@ public class Weapon {
   /**
    * Returns the weapon's name
    * */
-  private String getName() {
+  public String getName() {
     return name;
   }
 
@@ -48,23 +47,5 @@ public class Weapon {
     return weight;
   }
 
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (!(o instanceof Weapon)) {
-      return false;
-    }
-    final Weapon weapon = (Weapon) o;
-    return getDamage() == weapon.getDamage() &&
-        getWeight() == weapon.getWeight() &&
-        getName().equals(weapon.getName()) &&
-        getType() == weapon.getType();
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(getName(), getDamage(), getWeight(), getType());
-  }
+  public abstract WeaponType getType();
 }