+++ src/test/java/com/github/cc3002/finalreality/model/CharacterTest/PlayerTest/IPlayerTest.java
@@ -1,0 +1,18 @@
+package com.github.cc3002.finalreality.model.CharacterTest.PlayerTest;
+
+public interface IPlayerTest {
+
+    void setName();
+    void setDefense();
+    void setHealthPoints();
+    void setPlayer();
+    void setClass();
+
+    void testName();
+    void testDefense();
+    void testHealthPoints();
+    void testClass();
+    void testWeapon();
+    void waitTurnTest();
+
+}