+++ src/test/java/com/github/cc3002/finalreality/model/WeaponTest/CommonWeaponsTests/KnifeTest.java
@@ -1,0 +1,11 @@
+package com.github.cc3002.finalreality.model.WeaponTest.CommonWeaponsTests;
+
+import com.github.abraham054.finalreality.model.weapon.WeaponTypes.CommonWeaponType;
+
+public class KnifeTest extends CommonWeaponTest{
+
+    @Override
+    public void setType() {
+        type = CommonWeaponType.KNIFE;
+    }
+}