+++ src/test/java/com/github/cc3002/finalreality/model/WeaponTest/IWeaponTest.java
@@ -1,0 +1,14 @@
+package com.github.cc3002.finalreality.model.WeaponTest;
+
+public interface IWeaponTest {
+    void setName();
+    void setWeight();
+    void setDamage();
+    void setType();
+    void setTestedWeapon();
+
+    void testName();
+    void testWeight();
+    void testDamage();
+    void testType();
+}